synchro: add Node and Launcher types
Adds a `Launcher` type which contains much of the logic of `libra-node`,
but modified and customized to support generic `VMExecutor`/`VMVerifier`
types (probably best leveraged as a single type that impls both).

This seems like a good potential use case for Abscissa components, and
perhaps all of the `Launcher::start_*` methods can eventually be
refactored into components that wrap the relevant Libra types, but for
now it's just one big type with a bunch of methods that start things up
in the "right" order (hopefully).